Update: Use only one date-range component
