--- packages/reports/app/components/filter-values/dimension-date-range.js
@@ -1,1 +1,0 @@
-export { default } from 'navi-reports/components/filter-values/dimension-date-range';
